### 2016-10-27 #82 @sckott
ccafs - Client for CCAFS General Circulation Models Data
An interface to data from Climate Change, Agriculture, and Food Security (CCAFS) General Circulation Models (GCM) data. Data is stored in Amazon S3, from which we provide functions to fetch data.

```
Package: ccafs
Title: Client for 'CCAFS' 'GCM' Data
Description: Client for Climate Change, Agriculture, and Food Security ('CCAFS')
    General Circulation Models ('GCM') data. Data is stored in Amazon 'S3', from
    which we provide functions to fetch data.
Version: 0.0.8.9100
Authors@R: person("Scott", "Chamberlain", role = c("aut", "cre"),
    email = "myrmecocystus@gmail.com")
    rappdirs (>= 0.3.1),
    raster (>= 2.5-8),
    tibble (>= 1.2),
    xml2 (>= 1.0.0),
    data.table (>= 1.9.6)
    roxygen2 (>= 5.0.1),
    covr,
    knitr

```

https://github.com/ropenscilabs/ccafs

Scientists, particularly those studying climate change, forecasting crop production, etc.

Noppers.

  - (_Do not submit your package separately to JOSS_)
