### 2016-11-05 #82 @noamross
### Editor checks:

---

#### Editor comments

Currently seeking reviewers.

I find the README and documentation a little sparse. I try to follow the rule-of-thumb that a user may encounter the data and APIs we wrap for the first time through the package help or README.  Package-level documentation should at least briefly describe the data source and format, and provide links to more detailed documentation at the source.

Below is `goodpractice::gp()` and `covr::package_coverage()` output.  I note that much missing test coverage (when run with `NOT_CRAN="true") is for caching functions.

```
── GP ccafs ──────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 50% of code
    lines are covered by test cases.

    R/cache_data.R:5:NA
    R/cache_data.R:18:NA
    R/caching.R:48:NA
    R/caching.R:49:NA
    R/caching.R:55:NA
    ... and 45 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor
    windows that are about 80 characters wide. Try make your lines shorter than 80
    characters

    R/caching.R:74:1
    R/cc_data_fetch.R:46:1
    R/cc_data_fetch.R:51:1
    R/cc_list_keys.R:22:1
    R/cc_list_keys.R:24:1
    ... and 2 more lines

────────────────────────────────────────────
ccafs Coverage: 50.50%
R/caching.R: 0.00%
R/cc_data_fetch.R: 23.08%
R/cc_data_read.R: 70.00%
R/cache_data.R: 84.62%
R/zzz.R: 90.48%
R/cc_list_keys.R: 100.00%
```
---

Reviewers:
Due date:
