### 2016-11-09 #82 @noamross
Reviewers: @mikoontz @manuramon
Due date: 30-11-2016
