### 2016-11-24 #82 @manuramon
Here is my review. I hope that helps.

---

The CCAFS package has been developed to have a first sight of Climate Change, Agriculture, and Food Security (CCAFS) General Circulation Models (GCM) data in a fast and easy manner. This initial version is written to gather data from CCAFS data from Amazon S3, but I think it would be of great interest to make it more general allowing the use of data stored in other servers/places (for instance, including an URL argument in cc_data_flech).

Anyway, the package is well designed and fits its purpose.

Here are some specific comments:

- Installation goes without problems in windows, linux and mac OS.
- A broader description about what the package does in README would be desirable.
- Description of functions are in some occasions a bit short but precise.
- In my system, vignette doesn’t load when calling `vignette()` function, although it is available in the package directory.
- The Vignette covers almost all functions but is quite short. I would like to have more information about what each function does and more examples. For instance, in the `cc_data_fetch` documentation, there is an example on how to subset zip files using the `cc_list_keys` and `grep` functions that I think is quite useful and it could be included in the vignette.
- There is a typo in `?cc_data_read` (line 1, Reaa instead of Read)
- There is a typo in `cc_list_keys`, at marker arguments: you put pecifies instead of specifies. Also, reference to `GET` function could include the name of the package that function belongs to (`httr`package), to make easier to search for information about that function. The same could apply to all the functions referred in the documentation and that have been imported from other packages (mainly `httr`).
- When running `cc_data_fetch` function, progress bar doesn’t ends in a newline, so the console prompt comes just after del 100% of the progress bar.
- If you run the examples in `cc_data_fetch` and `cc_data_read` documentation without running first the example on README, `plot()` doesn’t work as the `raster` library is not loaded. I suggest to add library(raster) in all examples containing a plotting function or perhaps, to load the `raster` library by default (when loading the package).
- As a suggestion, the use of the `rasterVis`package instead of `raster will give you more control over the graphical outputs
