### 2016-11-30 #82 @mikoontz
>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly staing problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

### Overview

The `ccafs` package provides an `R` interface to geospatial data from the Climate Change Agriculture and Food Security General Circulation Models. The package is streamlined and works well with available geospatial frameworks in `R`, namely the `raster` package. The fetch-read workflow works nicely, and I appreciate the caching functionality to speed up data retrieval since these kinds of data can be quite large. Just as importantly, the cache cleaning operations are intuitive and convenient and will prove valuable to those using these data a lot. The Climate Change Agriculture and Food Security website (unaffiliated with the package) is well-organized but if I knew what data I needed to obtain from it, I would certainly prefer to use the `ccafs` package to do so.

I have one broad question and some more minor points. Depending on the outcome of the conversation around that broad question, many of these points can probably be addressed with modifications and/or additions to the documentation.

I was unable to get automatic checks from `goodpractice` or `covr` to run properly, but I can dig deeper to find out how I'm supposed to do that if adding to Noam's use of those tools would be helpful. (This is why I didn't check **Automated tests** above)

### Broad Question

I am having trouble envisioning how the `ccafs` package is intended to fit in the target audiences' workflow with CCAFS data. I can see two possible scenarios:

1. The package is designed to be used in tandem with the website. For instance, use the website to find a particular key of interest, use `ccafs` to get data for that particular key, then proceed with analysis in `R`.
2. The package is designed to be a stand-alone interface to the CCAFS data. For instance, a user can load the `ccafs` package, learn about what data are available, filter by region or data type at will, and download corresponding data.

In either scenario, I think the package documentation would benefit from a minimum description of the data available and how it is structured.

If scenario 1 is the intended use, I think the documentation could also benefit by pointing users to the CCAFS website where more can be learned about how to find particular data of interest. I noticed on the [ccafs-climate.org](ccafs-climate.org) website that there are ways of filtering what particular tiles or datasets can be downloaded. At a minimum, it would help a new user of the `ccafs` package to have a link to the website and how to find the 'key' information after searching, via the website interface, for the data they want.

If it's scenario 2, I think the package would benefit from more direction as to what keys a user might be interested in. This might require additional code to help translate user-desired filtering criteria into key filtering criteria (regular expressions of some kind?). If this functionality *is* available via the `prefix=`, `marker=`, and `delimiter=` arguments in the `cc_list_keys()` function, then I think the package would benefit from more explanation on how to use them to find particular datasets.

In scenario 2, I think the documentation wouldn't necessarily need to be augmented beyond what would be necessary for scenario 1.

### Minor Points

1. Can the raster package be loaded as a dependency upon loading the ccafs package? Is there any reason not to do this, if the `ccafs` packages' functionality centers around raster objects? Is it fair to say the keys don't do much until the raster data for those keys are retrieved? Or is there also a benefit to being able to readily acquire the files, even if they aren't read into `R`?

2. Relatedly, I see that the `raster()` and `brick()` functions are used for the `cc_data_read.character` method, which can happen even without an explicit call to `library(raster)`. The `plot()` function won't work without an explicit `library(raster)` call first. I'm not familiar with how that works, or what the advantage is. What is the benefit of letting the `cc_data_read` method use the `raster` package behind the scenes, but not the `plot` method?

3. The package help (`?ccafs`) would benefit from some more description. Where is the original data source? What does CCAFS GCM stand for? Some of this information is in the vignette and on the GitHub page, and I think it would help a new user to have it available here, too.

4. I see that the CCAFS data has an attribution/non-commercial license. Is there a need to pass along the fact that there is a "non-commercial" part of the data license to the user of the package? What about passing along the citation information for the data found [here](http://ccafs-climate.org/about/)? Can this be done in the documentation or in the

5. Can the vignette be linked to in the `?ccafs`? Or be made available with a call to `vignette("vignette_name")`?

6. I got the below warnings when trying to run `citation("ccafs")`. Is the lack of a date in the DESCRIPTION file just because the package isn't published yet?

```
To cite package ‘ccafs’ in publications use:

  Scott Chamberlain (NA). ccafs: Client for 'CCAFS' 'GCM' Data. R package version 0.0.8.9100.
  https://github.com/ropenscilabs/ccafs

A BibTeX entry for LaTeX users is

  @Manual{,
    title = {ccafs: Client for 'CCAFS' 'GCM' Data},
    author = {Scott Chamberlain},
    note = {R package version 0.0.8.9100},
  }

1: In citation("ccafs") :
  no date field in DESCRIPTION file of package ‘ccafs’
2: In citation("ccafs") :
  could not determine year for ‘ccafs’ from package DESCRIPTION file
```

7. I see the term 'bucket' referenced a couple of times and I think I can infer its meaning from context within the function definitions, but perhaps this is something that could be mentioned in the introduction to the package if it is an important part of the CCAFS file organization structure.

8. `cc_data_fetch()` and `cc_data_read()` examples need a `library(raster)` call prior to call to `plot()` in order for them to run. This is the reason why I didn't check the box next to **Examples** above.

9. When `cc_cache_details()` is run on an NA (for instance, a file that doesn't exist), no error or warning is returned-- just an NA detail for the size. Is this expected behavior, or should a warning/error be reported as when files don't exist on a call to `cc_cache_delete()`?

```
> cc_cache_details(files = "foo")
<ccafs cached files>
  directory: /Users/mikoontz/Library/Caches/ccafs

  file: foo
  size: NA mb
```

10. In the vignette, can it be made more clear in the following code chunk that the `sub()` function is piecing the pasted character object into a single, long URL. I worry that users unfamiliar with regular expressions may have trouble understanding why this key is written this way.

```
key <- sub("\n|\\s+", "", paste0(
  "ccafs/ccafs-climate/data/ipcc_5ar_ciat_downscaled/rcp2_6/
  2030s/bcc_csm1_1_m/10min/",
  "bcc_csm1_1_m_rcp2_6_2030s_prec_10min_r1i1p1_no_tile_asc.zip"))
```

11. Is there a different purpose for `[` compared to `[[`? Is the only difference that `[` lets a user access multiple ccafs_files objects and doesn't drop names? It looks like the class remains the same for each. Does defining `[[` just ensure that a user doesn't break their code by trying to subset this way?

12. Line 9 in cc_list_keys.R: "pecifies" should be "specifies"
