### 2016-11-30 #82 @noamross
Thank you for your reviews, @mikoontz and @manuramon!

@sckott When responding to Mike's point, I think we should consider the evolving standard of "the package may be the users' first encounter with this data/API".  This doesn't mean one has to document everything in the package, but the README, vignettes, and other documentation should point the reader to adequate online documentation if necessary.

I also suggest that one possible solution to the dealing with the **raster** dependencies is to re-export essential functions and methods from that package (for print/plot etc.), rather than attaching its whole namespace via `Depends:`. In the docs/vignettes/README you'll want to point out that these are raster objects and one should load **raster** for further processing.

I will take a closer look at the testing suite in a bit, as well.  Mike, in general the Automated Tests category doesn't refer to the tests we run, but the package tests housed in `tests/testthat`. The test badges at the top of the README will lead you to the results if you are unable to get them to run on your computer (with `devtools::test()`)
