### 2016-12-01 #82 @mikoontz
Oh great, thanks for that. In that case, I can update my review:

All tests in `tests/testthat` pass on my local machine. The file `test-all.R` in the `tests/` directory gives the below error. Is that expected?

```
> library("testthat")
> test_check("ccafs")
Error: No tests found for ccafs
```

The badges suggest that the automatic tests by AppVeyor and Travis-CI work. I'm not clear on whether reporting that fulfills my reviewer duties. I'm not sure how to get `devtools::test()` to work on my machine.

My intuition is to pass the `test()` function the path for my `ccafs` package:

```
devtools::test(pkg = paste0(.libPaths(), "/ccafs"))
```

But I can't image that's correct, since it returns a bunch of failed tests that worked fine when I copied the code directly from within the `tests/testthat` directory to my R console.

```
> devtools::test(pkg = paste0(.libPaths(), "/ccafs"))
Loading ccafs
Testing ccafs
cc_data_fetch: 123
cc_data_read: 456
cc_list_keys: 789a

Failed -------------------------------------------------------------------------------------------------------------------
1. Error: cc_data_fetch works (@test-cc_data_fetch.R#8) ------------------------------------------------------------------
could not find function "cc_data_fetch"
1: .handleSimpleError(function (e)
   {
       e$call <- sys.calls()[(frame + 11):(sys.nframe() - 2)]
       register_expectation(e, frame + 11, sys.nframe() - 2)
       signalCondition(e)
   }, "could not find function \"cc_data_fetch\"", quote(eval(expr, envir, enclos))) at /Library/Frameworks/R.framework/Versions/3.3/Resources/library/ccafs/tests/testthat/test-cc_data_fetch.R:8
2: eval(expr, envir, enclos)

...
```
