### 2016-12-04 #82 @noamross
@mikoontz `devtools::test()` and other tools like `covr`work on the project development directory, not the installed package, set the cloned git repo as your R working directory, and run everything from there.

@sckott I took a look at your testing and see there isn't coverage for stuff in `cacheing.R`.  These are mostly status print functions so I think this OK.  However, it makes me consider what is appropriate testing for the cache functionality (something I'm dealing with myself at the moment).  I think the approriate test is to determine whether the cached files really do match what would be downloaded.  Running such tests would protect against code changes that result in changes to the cached paths, or unexpected changes to the files at the data source.  An approach to this would be having an artificial cache directory for testing (perhaps stashed somewhere else convenient, so as not to clutter the repo), and test that downloaded files match what's there.
I've taken a look at tests and my main thought is that test
