### 2016-12-04 #82 @sckott
> consider the evolving standard of "the package may be the users' first encounter with this data/API". This doesn't mean one has to document everything in the package, but the README, vignettes, and other documentation should point the reader to adequate online documentation if necessary.

Good idea.

> I also suggest that one possible solution to the dealing with the raster dependencies is to re-export essential functions and methods from that package (for print/plot etc.)

makes sense

> However, it makes me consider what is appropriate testing for the cache functionality (something I'm dealing with myself at the moment).

good ideas, will have a go at adding cache testing
