### 2016-12-15 #82 @sckott
responses to @manuramon

------

> I think it would be of great interest to make it more general allowing the use of data stored in other servers/places (for instance, including an URL argument in cc_data_flech).

Hmm, thanks for the idea, but i think it's important to keep this pkg specific to CCAFS data, and make new pkgs (following similar pattern) for other data sources.

specific comments:

> - A broader description about what the package does in README would be desirable.

Okay, can do!

> - Description of functions are in some occasions a bit short but precise.

Thanks! Will expand on what each function does.

> - In my system, vignette doesn’t load when calling `vignette()` function, although it is available in the package directory.

i guess `vignette(package = "ccafs")` brings up a man page for the vignette, and `vignette("ccafs_vignette", package = "ccafs")` opens the vignette. Does that do what you think should be done?

> - The Vignette covers almost all functions but is quite short. I would like to have more information about what each function does and more examples. For instance, in the `cc_data_fetch` documentation, there is an example on how to subset zip files using the `cc_list_keys` and `grep` functions that I think is quite useful and it could be included in the vignette.

Good point, will add that specific thing, and will fill out the vignette more.

- There is a typo in `?cc_data_read` (line 1, Reaa instead of Read)

will fix it. thanks!

- There is a typo in `cc_list_keys`, at marker arguments: you put pecifies instead of specifies. Also, reference to `GET` function could include the name of the package that function belongs to (`httr`package), to make easier to search for information about that function. The same could apply to all the functions referred in the documentation and that have been imported from other packages (mainly `httr`).

will fix the type. thanks! for reference to `GET` are you talking about the param definition for `...`? if so, the syntax behind that is `\code{\link[httr]{GET}}` which creates a link to the man file for GET, does that seem adequate?

- When running `cc_data_fetch` function, progress bar doesn’t ends in a newline, so the console prompt comes just after del 100% of the progress bar.

good catch. hmm, that's done via `httr::progress` - i'm not sure i can change that behavior but I can  probably change the print behavior for the output class

- If you run the examples in `cc_data_fetch` and `cc_data_read` documentation without running first the example on README, `plot()` doesn’t work as the `raster` library is not loaded. I suggest to add library(raster) in all examples containing a plotting function or perhaps, to load the `raster` library by default (when loading the package).

thanks for catching that, I'll just add `library(raster)` to the example

- As a suggestion, the use of the `rasterVis`package instead of `raster` will give you more control over the graphical outputs

Nice, thanks, will try that out, at least for possibly including for examples.

---

I'll open issues and link those to above comments as appropriate after looking at the other review
