### 2016-12-15 #82 @sckott
responses to @mikoontz

------

> Broad Question

by `the website` I assume you mean http://ccafs-climate.org ?

I'm not sure what is best. Do you have a preference as a (i assume) potential user?

> 1. Can the raster package be loaded as a dependency upon loading the ccafs package? Is there any reason not to do this, if the `ccafs` packages' functionality centers around raster objects? Is it fair to say the keys don't do much until the raster data for those keys are retrieved? Or is there also a benefit to being able to readily acquire the files, even if they aren't read into `R`?

as @noamross said i'd prefer to re-export essential fxns over putting in `Depends`. It is fair to say the keys don't do much, they're just character strings pointing at a location on amazon s3. I imagine there is a use case for downloading the files and reading in a later R session, so i don't think we should assume the user always needs `raster` loaded by default.

> 2. Relatedly, I see that the `raster()` and `brick()` functions are used for the `cc_data_read.character` method, which can happen even without an explicit call to `library(raster)`. The `plot()` function won't work without an explicit `library(raster)` call first. I'm not familiar with how that works, or what the advantage is. What is the benefit of letting the `cc_data_read` method use the `raster` package behind the scenes, but not the `plot` method?

The reasoning behind that is to keep the package as light weight as possible, e.g., https://github.com/ropenscilabs/ccafs/blob/master/NAMESPACE#L23-L24 shows we only import two of the `raster` functions. WE can however import and export `raster::plot` - which I may do, i'll play with that first and see how it works out

> 3. The package help (`?ccafs`) would benefit from some more description. Where is the original data source? What does CCAFS GCM stand for? Some of this information is in the vignette and on the GitHub page, and I think it would help a new user to have it available here, too.

Thanks! Will definitely improve the pkg level file

> 4. I see that the CCAFS data has an attribution/non-commercial license. Is there a need to pass along the fact that there is a "non-commercial" part of the data license to the user of the package? What about passing along the citation information for the data found [here](http://ccafs-climate.org/about/)? Can this be done in the documentation or in a call to some function?

No rule per se for CRAN but yeah, good idea, i'll add that in the pkg level man file and in readme

> 5. Can the vignette be linked to in the `?ccafs`? Or be made available with a call to `vignette("vignette_name")`?

it is available like `vignette(package = "ccafs")` or `vignette("ccafs_vignette", package = "ccafs")`, but i can add that to the pkg level man file

> 6. I got the below warnings when trying to run `citation("ccafs")`. Is the lack of a date in the DESCRIPTION file just because the package isn't published yet?

It seems to be common now to not put `Date` in DESCRIPTION file, as when CRAN builds the pkg they add that in for the date built on. It's hard to remember to update that Date field, so another reason not to include it

> 7. I see the term 'bucket' referenced a couple of times and I think I can infer its meaning from context within the function definitions, but perhaps this is something that could be mentioned in the introduction to the package if it is an important part of the CCAFS file organization structure.

yeah, makes sense to add more explanation. a bucket is the top level thing that files are organized in within S3

> 8. `cc_data_fetch()` and `cc_data_read()` examples need a `library(raster)` call prior to call to `plot()` in order for them to run. This is the reason why I didn't check the box next to **Examples** above.

right, thanks! noted by other reviewer as well see https://github.com/ropenscilabs/ccafs/issues/5

> 9. When `cc_cache_details()` is run on an NA (for instance, a file that doesn't exist), no error or warning is returned-- just an NA detail for the size. Is this expected behavior, or should a warning/error be reported as when files don't exist on a call to `cc_cache_delete()`?

Probably not the expected behavior. Perhaps returning a `file doesn't exist` or similar message would be useful. though don't want to stop with error as user could pass in some files that do exist, and some that don't

> 10. In the vignette, can it be made more clear in the following code chunk that the `sub()` function is piecing the pasted character object into a single, long URL. I worry that users unfamiliar with regular expressions may have trouble understanding why this key is written this way.

Right, should include a simple explanation about that. When the pkg manual is made they don't wrap lines, so that's why i try to make sure all my code and docs are 80 line width or less. Although in the vignette its md/html, so I guess doesn't matter there so much.

> 11. Is there a different purpose for `[` compared to `[[`? Is the only difference that `[` lets a user access multiple ccafs_files objects and doesn't drop names? It looks like the class remains the same for each. Does defining `[[` just ensure that a user doesn't break their code by trying to subset this way?

Yeah, you got it. The single bracket method is for indexing like `[1:3]`, while the double bracket method  is for indexing like `[[1]]` - and yeah, the goal with those is that the S3 class isn't lost, since the actual data within the object is just a character vector of file paths. try `unclass()` on the output of `cc_data_fetch()`

> 12. Line 9 in cc_list_keys.R: "pecifies" should be "specifies"

thanks, will fix

------

I'll open issues and link those to above comments as appropriate after looking at the other review
