### 2017-01-03 #82 @mikoontz
I'll just respond to the points where you have a question for me. Let me know what other feedback would be helpful.

> by the website I assume you mean [http://ccafs-climate.org](http://ccafs-climate.org)?
> I'm not sure what is best. Do you have a preference as a (i assume) potential user?

Whoops, yes I meant the [http://ccafs-climate.org](http://ccafs-climate.org) website.

I would be comfortable working with both platforms  in parallel (your package in conjunction with the official CCAFS website) as long as there is some more direction from your package about how to leverage the official websites search capabilities to find the key strings that would be necessary. The package would still provide a valuable service to a novice user. I don't think there is a need to build in a spatial/temporal/data type subsetting engine directly into the package.

> as @noamross said i'd prefer to re-export essential fxns over putting in Depends. It is fair to say the keys don't do much, they're just character strings pointing at a location on amazon s3. I imagine there is a use case for downloading the files and reading in a later R session, so i don't think we should assume the user always needs raster loaded by default.

Re-exporting essential functions sounds great! I didn't know that was possible. I am on board with keeping the package lightweight, and allowing for potential use cases of the keys that may not require the entire `raster` package.
