### 2017-01-03 #82 @sckott
@mikoontz thanks for your feedback!

I would like to add functionality for searching for what data is available, but they sure aren't making it easy to do programmatically - will keep looking into that

re-exporting it is
