### 2017-01-04 ropenscilabs/ccafs#12 @sckott
review: re-export essential raster fxns in pkg
comments near top of https://github.com/ropensci/onboarding/issues/82#issuecomment-267250659
