### 2017-01-05 #82 @sckott
Issues linked in my responses above - dealing with those now
