### 2017-01-06 #82 @sckott
All issues raised by reviewers have been addressed - Anything else to do?

There still isn't a way to search for data - for a future package version,  I'll see if I can scrape the search page for CCAFS, but i don't think that's a good long-term plan.
