### 2017-01-06 #82 @noamross
Editor comments have been addressed.  @mikoontz and @manuramon, could you please take a look and let us know that @sckott's changes have addressed your concerns?
