### 2017-01-07 #82 @manuramon
Hi @sckott @noamross and @mikoontz. For me, @sckott has addressed all our concerns in a pretty way. I think the resulted package is nice and useful and for sure that in a (not so far) future this package will be improved with new functionalities (such us search for data).

PS: the example in the cc_data_fetch using the progress bar results in one line printed for each percentage point in the vignette. This is a not important issue (in fact, there is not an issue)
