### 2017-01-09 #82 @mikoontz
Hi @sckott @noamross and @manuramon,

I think the package is much improved, largely thanks to the expanded documentation. Thanks for the extra information!

I have 4 concerns still:

*First*, I can't get the vignette to run using either of these lines of code:

```
> vignette(package = "ccafs")
no vignettes found
> vignette("ccafs_vignette", package = "ccafs")
Warning message:
vignette ‘ccafs_vignette’ not found
```

*Second*, I still think that the vignette could use an example workflow to get the key from the website, even given the current state of the search functionality with the R interface (i.e. not present, but perhaps in development).

As an example, I used the [CCAFS website](http://ccafs-climate.org/) to try to get a key like a user might. Specifically, I tried to recreate getting the key that the vignette uses.  Here was my workflow:

1. Navigate to [Data -> Spatial Downscaling](http://ccafs-climate.org/data_spatial_downscaling/)
1. On the left bar -> File Set, I clicked "Delta Method IPCC AR5" under the Empirical/Statistical Downscaling
1. On the left bar -> Scenario, I clicked "RCP 2.6"
1. On the left bar -> Model, I clicked the "bcc_csm1_1_m"
1. On the top bar -> Extent, I clicked "Global"
1. On the top bar -> Format, I clicked "ASCII Grid Format"
1. On the top bar -> Period, I clicked "2030s"
1. On the top bar -> Variable, I clicked "Precipitation"
1. On the top bar -> Resolution, I clicked "10 minutes"
1. This resulted in 1 file found, and I clicked "Search"
1. I checked the box next to the file returned by the search, and clicked "Generate Download Links"
1. I clicked "Skip" in the paragraph to go to the download links directly without submitting my email or basic information which would be used by CCAFS for impact assessment. (side note: this prompted concern #4, see below)
1. I copied the address of the file to download (Right click -> Copy Link Address)
1. I pasted the link address into my R script, and deleted everything before "ccafs/ccafs-climate/data..." and used that truncated character string as my key in the cc_data_fetch() call.

Is this the workflow that you would use? If so, can you include it with the vignette or is that outside the scope of the vignette?

*Third*, is this package unable to get the regional datasets? I tried to get the region A1 version of the above key (with a 30 second resolution, since that's all that is available) and the package can fetch but not read. I tried a few other combinations of regional datasets and had the same problem. Are there other data sets you know aren't retrievable? I couldn't figure out how to get a key from the "Bias Correction" or the "Weather Station" parts of the website (I could only retrieve data from the "Spatial Downscaling" section); does that mean those data can't be accessed by the `ccafs` package?

```
# Example of a regional dataset that I couldn't get to work with cc_data_read()
key7 <- "ccafs/ccafs-climate/data/ipcc_5ar_ciat_tiled/rcp2_6/2030s/bcc_csm1_1_m/30s/bcc_csm1_1_m_rcp2_6_2030s_prec_30s_r1i1p1_a1_asc.zip"
(test7 <- cc_data_fetch(key = key7))
cc_data_read(test7) # Doesn't read the file
```

When I run `cc_cache_list()` I see a similar set of files as I do when I use the key from the vignette except each .asc seems to have a corresponding .prj associated with it. Maybe this is what prevents the rest of the `ccafs` functions from working on those files?

*Fourth*, have you had any communication with the CCAFS/CGIAR folks about accessing their data via R instead of their interface? From their "Terms and Conditions" paragraph (copied below) that pops up before you can get a download link, it seems like they'd be interested in assessing the impact of their work.

> To continue downloading your files, please first fill in your email and then some basic information. This information will be used by CCAFS solely for impact assessment and CGIAR and Center level reporting purposes. Filling it in will greatly help us to track the use of the portal and keep improving it. This portal provides data to a very large community of users and improving its usability and efficiency is a key aspect we work on continuously. However, you may click on skip to download links directly.
