### 2017-01-09 #82 @manuramon
@mikoontz to get the vignette working I run the `devtools::install_github()` function with the option `build_vignettes = TRUE`.

I also have tried to get a regional data set (the example @mikoontz pointed out above) and received the following error (_Error in .rasterObjectFromFile(x, band=... : Cannot create a RasterLayer from this file_) when running `cc_data_read`. However, if I read the .asc file from the cache directory (`raster::raster("~/Library/Caches/.../prec_1.asc")`) I didn't get any error and I obtained the plot.
