### 2017-01-09 #82 @mikoontz
Thanks @manuramon, you're right. I can run the vignettes now using the `build_vignettes = TRUE` argument in the `install_github()` function. I assume they'll be loaded automatically when the package goes on CRAN, and so this isn't a concern anymore.

I also see that I can call the `raster()` function directly on the .asc file in the cache and get the desired plot. The contents from the .prj file are read automatically with their associated .asc if they are in the same directory (see [here](http://gis.stackexchange.com/questions/138256/loading-projection-along-with-raster-in-r)).

```
# .asc file in same directory as the associated .prj
> test <- raster("/Users/mikoontz/Library/Caches/ccafs/.../prec_1.asc")
> crs(test)
CRS arguments:
 +proj=longlat +ellps=WGS84 +towgs84=0,0,0,0,0,0,0 +no_defs

# .asc file moved to a different location to decouple it with the .prj
> test2 <- raster("/Users/mikoontz/Desktop/prec_1.asc")
> crs(test2)
CRS arguments: NA
```
