### 2017-01-09 #82 @sckott
> the example in the cc_data_fetch using the progress bar results in one line printed for each percentage point in the vignette

right - it can be surpressed, which I'll do for the vignette probably
