### 2017-01-09 #82 @sckott
thanks for the help on vignette @manuramon

@mikoontz

2. Yes, I'll include in the vignette or in another a workflow for getting keys in the browser, good idea. To be honest, I hadn't really used the web interface to drill down to data since I was focused on the data coming from the S3 source

3. regional data sets and such ...

I don't know, I'll have a look

4. I haven't been in contact. The best thing I could do IMO is put in the user-agent string the package name (`ccafs`) and version, and that it's coming from R.  They probably collect logs (or they should if they dont) on the downloads from Amazon, so they'll see the user-agent string and get a sense for how many people are using this pkg

>  I assume they'll be loaded automatically when the package goes on CRAN, and so this isn't a concern anymore.

Yes.
