### 2017-01-09 ropenscilabs/ccafs#18 @sckott
fix problem in reading asc files that have associated proj files
via https://github.com/ropensci/onboarding/issues/82#issuecomment-271205656
