### 2017-01-09 #82 @sckott
AFAICT "Bias Correction" or the "Weather Station"  sections are somehow private data for people that have access.
