### 2017-01-09 #82 @sckott
@mikoontz  Try again after reinstalling `devtools::install_github("ropenscilabs/ccafs")`  - Using `raster::stack` now instead of `raster::brick`, which didn't support using PROJ files  (stack does). That should allow you to read that key you had in the eg. And like in your comment above, `raster::raster()` also supports using the PROJ file.
