### 2017-01-10 #82 @sckott
if you two have time @manuramon @mikoontz try out new function `cc_search` after reinstalling `devtools::install_github("ropenscilabs/ccafs")` not going to work for all query combinations yet, and some errors for sure, but examples should work
