### 2017-01-11 #82 @manuramon
@sckott I have found the `cc_search` function useful. I find the numerical recoding difficult to remember and it forces you to go through the help, but I can't imagine a better way to do it. The example in the `cc_search` function help using the `lapply` function is very useful.
