### 2017-01-11 #82 @sckott
@manuramon thanks, agree that there's not a better way to do it, for the few parameters that only have two options i just use a character string, but most parameters have at least 6 options many of which are quite long strings with spaces and such,
