### 2017-01-18 #82 @mikoontz
**Vignette**

Can you add the "amazon_s3_keys" vignette to the package help `?ccafs` page?

```
vignette("amazon_s3_keys", package = "ccafs")
```

** Regional Datasets **

I'm now able to read the regional datasets that incorporate their associated .prj files. It looks like the ccafs_files object print method isn't reporting the correct number of files. When I read the same regional key that we've been using, the resulting object from the `cc_data_fetch()` function call only shows 1 file, and doesn't show the file type. Twelve plots are made after reading the fetched object, so I think it's just an accounting error (possibly also because of the extra .prj files in the cache?).

```
# Example of a regional dataset that I couldn't get to work with cc_data_read() in the last iteration
key7 <- "ccafs/ccafs-climate/data/ipcc_5ar_ciat_tiled/rcp2_6/2030s/bcc_csm1_1_m/30s/bcc_csm1_1_m_rcp2_6_2030s_prec_30s_r1i1p1_a1_asc.zip"
(test7 <- cc_data_fetch(key = key7))
test7 # Aren't there supposed to be 12 files here?

<CCAFS GCM files>
   1 files
   Base dir: /bcc_csm1_1_m_rcp2_6_2030s_prec_30s_r1i1p1_a1_asc
   File types (count):
>
```

** `cc_search()` function **

The `cc_search()` function is awesome. Thanks for adding that! It looks like you access the CCAFS file-list.php file directly. Clever!

Just 2 comments on it:

1. The exhaustive list of possible argument values in `ccafs-search` is great. I think it would be even more amazing if you could leverage some tab autocomplete power to help users quickly visualize the possible argument values. I was having a hard time finding an ability in R to tab complete the possible valid values of an argument in a function call-- I got the sense that capability doesn't exist. What about defining a named list of named lists object to translate human-readable text of the different valid argument values into their lookup codes? Something like this... (I didn't finish adding all of the model options, but the other categories are complete I think)

```
cc_params <- list(
  file_set = list("Delta method IPCC AR5" = 12,
                  "Delta method IPCC AR4" = 4,
                  "MarkSim Pattern Scaling" = 9,
                  "Eta South America" = 10,
                  "PRECIS Andes"= 7,
                  "CORDEX" = 8,
                  "Disaggregation IPCC AR4" = 11,
                  "Delta Climgen" = 3,
                  "Delta Method IPCC AR4 (Climgen Data)" = 2,
                  "Delta Method IPCC AR4 (Stanford Data)" = 5,
                  "Delta Method IPCC AR3" = 6),
  scenario = list("Baseline" = 1,
                  "SRES A1B" = 2,
                  "SRES A2A" = 3,
                  "SRES B2A" = 4,
                  "SRES A2" = 5,
                  "SRES B1" = 6,
                  "RCP 2.6" = 7,
                  "RCP 4.5" = 8,
                  "RCP 6.0" = 9,
                  "RCP 8.5" = 10),
  model = list("Baseline" = 1,
               "bcc_csm1_1" = 42,
               "bcc_csm1_1_m" = 43,
               "bccr_bcm2_0" = 2,
               "bnu_esm" = 44,
               "cccma_cancm4" = 45,
               "cccma_canesm2" = 46),
  extent = list("global" = "global",
                "region" = "region"),
  format = list("ascii" = 1,
                "esri" = 2),
  period = list("1970s" = 1,
                "1990s" = 10,
                "2000s" = 2,
                "2020s" = 3,
                "2030s" = 4,
                "2040s" = 5,
                "2050s" = 6,
                "2060s" = 7,
                "2070s" = 8,
                "2080s" = 9),
  variable = list("Bioclimatics" = 1,
                  "Diurnal Temperature Range" = 6,
                  "Maximum Temperature" = 3,
                  "Mean Temperature" = 4,
                  "Minimum Temperature" = 5,
                  "Precipitation" = 2,
                  "Solar Radiation" = 7,
                  "Other" = 9999),
  resolution = list("30 seconds" = 1,
                    "2.5 minutes" = 2,
                    "5 minutes" = 3,
                    "10 minutes" = 4,
                    "30 minutes" = 5,
                    "25 minutes" = 6,
                    "20 minutes" = 7),
  tile = list("A1" = "A1",
              "A2" = "A2",
              "A3" = "A3",
              "A4" = "A4",
              "A5" = "A5",
              "A6" = "A6",
              "B1" = "B1",
              "B2" = "B2",
              "B3" = "B3",
              "B4" = "B4",
              "B5" = "B5",
              "B6" = "B6",
              "C1" = "C1",
              "C2" = "C2",
              "C3" = "C3",
              "C4" = "C4",
              "C5" = "C5",
              "C6" = "C6"))
```

An example use case would turn the directly coded search of the file key like you had in the `?cc_search` example (reproduced below)...

```
(res <- cc_search(file_set = 4, scenario = 6, model = 2, extent = "global", format = "ascii", period = 5, variable = 2, resolution = 3))
```

...into something like the code below, where tab complete helped every step of the way.

```
(res <- cc_search(file_set = cc_params$file_set$`Delta method IPCC AR4`,
                 scenario = cc_params$scenario$`SRES B1`,
                 model = cc_params$model$bccr_bcm2_0,
                 extent = cc_params$extent$global,
                 format = cc_params$format$esri,
                 period = cc_params$period$`2040s`,
                 variable = cc_params$variable$Precipitation,
                 resolution = cc_params$resolution$`5 minutes`))
```

 I have no idea whether it is good practice to hard code a lookup table like this, but maybe there is a better way to implement the concept. It may also make it easier to for the package maintainer to update the `cc_search()` function if the CCAFS group changes their codes by updating this object rather than the help file. Maybe that'd save the end user from needing to double check the help file every time?

1. The cc_search() function returns errors if the format= or extent= arguments are integer values, but the `?cc_search` help ("Arguments" section) suggests each of these arguments should (or can?) take a value of 1 or 2.
