### 2017-01-18 #82 @sckott
> Can you add the "amazon_s3_keys" vignette to the package help ?ccafs page?

yeah, will do

I'll check on the regional datasets problem

## cc_search

* glad you like `cc_search()`
* pop up hints for parameters are there, but yeah kind of unwieldy when there's so many options. Good idea for a list of options. Right that it will have to be updated if they update their options, kind of annoying, but I can try that and see if it mkakes sense.
* added input type checkers for the parameters, so will fail well on wrong inputs now
