### 2017-02-14 #82 @sckott
@mikoontz Okay, I think I addressed all your points - changes made -

anything else I should do folks?
