### 2017-02-15 #82 @mikoontz
Awesome. I see all the points as addressed: vignette is in the help file, good accounting of the files coming from the regional keys, and a way to search with helpful prompts.

Green light from me! This is great!
