### 2017-02-15 #82 @manuramon
Great job @sckott. I think this final version of the `ccafs` package is easy to use very useful for the purpose it was developed.

Green light from me, as well.

PS. For me this has been the first time I have reviewed an R package and I have learned a lot and enjoyed it, so thanks to all of you! I hope to have the opportunity to review more packages in the future.
