### 2017-02-22 #82 @noamross
Thanks @mikoontz and @manuramon for your reviews and excellent follow-up! @sckott, I've done final checks and we're good to go.
